requirements.txt
Flask==0.12.2
Flask-SQLAlchemy==2.2
mysql-python==1.2.5
colorlog==4.0.2
pymysql==0.9.3
flask_mail==0.9.1
gevent==1.4.0
flask-restplus==0.12.0
flask-jwt-extended==3.18.0
requests==2.18.4
flask_login==0.4.0
flask_wtf==0.14.2